codex(web): extract archive analytics store
