+++ web/db_state.py
@@ -5,16 +5,8 @@ from __future__ import annotations
 import hashlib
 import json
 import os
-import re
 import sqlite3
-import uuid
-from datetime import datetime, timedelta, timezone
 from typing import Any, Dict, Optional, cast
-
-try:
-    from archive import build_archive_entry
-except ImportError:
-    from web.archive import build_archive_entry  # pragma: no cover
 
 try:
     import db_core as _db_core
@@ -25,6 +17,16 @@ try:
     import db_history as _db_history
 except ImportError:
     from web import db_history as _db_history  # pragma: no cover
+
+try:
+    import db_archive as _db_archive
+except ImportError:
+    from web import db_archive as _db_archive  # pragma: no cover
+
+try:
+    import db_analytics as _db_analytics
+except ImportError:
+    from web import db_analytics as _db_analytics  # pragma: no cover
 
 APPROVAL_STATUS_NOT_REQUESTED = _db_history.APPROVAL_STATUS_NOT_REQUESTED
 APPROVAL_STATUS_PENDING = _db_history.APPROVAL_STATUS_PENDING
@@ -66,6 +68,13 @@ update_history_status = _db_history.update_history_status
 get_history_row_by_idempotency_key = _db_history.get_history_row_by_idempotency_key
 get_history_row = _db_history.get_history_row
 update_history_review_state = _db_history.update_history_review_state
+sync_archive_entry_from_history = _db_archive.sync_archive_entry_from_history
+sync_archive_entries_from_history = _db_archive.sync_archive_entries_from_history
+get_archive_entry = _db_archive.get_archive_entry
+search_archive_entries = _db_archive.search_archive_entries
+record_analytics_event = _db_analytics.record_analytics_event
+list_analytics_events = _db_analytics.list_analytics_events
+get_analytics_dashboard_data = _db_analytics.get_analytics_dashboard_data
 
 
 def _connect(db_path: str) -> sqlite3.Connection:
@@ -466,506 +475,3 @@ def get_active_source_policies(db_path: str) -> Dict[str, list[str]]:
         return {"allowlist": allowlist, "blocklist": blocklist}
     finally:
         conn.close()
-
-
-def _parse_history_json(raw_value: str | None) -> Dict[str, Any]:
-    """Best-effort JSON decoding for stored history payloads."""
-    if not raw_value:
-        return {}
-    try:
-        parsed = json.loads(raw_value)
-    except json.JSONDecodeError:
-        return {}
-    return parsed if isinstance(parsed, dict) else {}
-
-
-def _serialize_event_timestamp(value: str | datetime | None) -> str:
-    """Serialize analytics timestamps in UTC ISO format."""
-    if value is None:
-        timestamp = datetime.now(timezone.utc)
-    elif isinstance(value, str):
-        return value
-    elif value.tzinfo is None:
-        timestamp = value.replace(tzinfo=timezone.utc)
-    else:
-        timestamp = value.astimezone(timezone.utc)
-
-    return timestamp.isoformat().replace("+00:00", "Z")
-
-
-def sync_archive_entry_from_history(
-    db_path: str, job_id: str
-) -> Optional[Dict[str, Any]]:
-    """Create or refresh a normalized archive entry from a completed history row."""
-    conn = _connect(db_path)
-    try:
-        cursor = conn.cursor()
-        cursor.execute(
-            """
-            SELECT id, params, result, created_at, status
-            FROM history
-            WHERE id = ?
-            """,
-            (job_id,),
-        )
-        row = cursor.fetchone()
-        if not row or row[4] != "completed":
-            return None
-
-        entry = build_archive_entry(
-            job_id=row[0],
-            created_at=row[3],
-            params=_parse_history_json(row[1]),
-            result=_parse_history_json(row[2]),
-        )
-        if entry is None:
-            return None
-
-        cursor.execute(
-            """
-            INSERT INTO archive_entries (
-                job_id,
-                title,
-                snippet,
-                source_kind,
-                source_value,
-                keywords,
-                metadata,
-                search_text,
-                created_at,
-                updated_at
-            )
-            VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?)
-            ON CONFLICT(job_id) DO UPDATE SET
-                title = excluded.title,
-                snippet = excluded.snippet,
-                source_kind = excluded.source_kind,
-                source_value = excluded.source_value,
-                keywords = excluded.keywords,
-                metadata = excluded.metadata,
-                search_text = excluded.search_text,
-                created_at = excluded.created_at,
-                updated_at = excluded.updated_at
-            """,
-            (
-                entry["job_id"],
-                entry["title"],
-                entry["snippet"],
-                entry["source_kind"],
-                entry["source_value"],
-                canonical_json({"items": entry["keywords"]}),
-                canonical_json(entry["metadata"]),
-                entry["search_text"],
-                entry["created_at"],
-                _serialize_event_timestamp(datetime.now(timezone.utc)),
-            ),
-        )
-        conn.commit()
-    finally:
-        conn.close()
-
-    return get_archive_entry(db_path, job_id)
-
-
-def sync_archive_entries_from_history(db_path: str, *, limit: int = 100) -> int:
-    """Backfill recent completed history rows into archive storage."""
-    conn = _connect(db_path)
-    try:
-        cursor = conn.cursor()
-        cursor.execute(
-            """
-            SELECT id
-            FROM history
-            WHERE status = 'completed'
-            ORDER BY created_at DESC
-            LIMIT ?
-            """,
-            (max(1, int(limit)),),
-        )
-        job_ids = [str(row[0]) for row in cursor.fetchall()]
-    finally:
-        conn.close()
-
-    synced = 0
-    for job_id in job_ids:
-        if sync_archive_entry_from_history(db_path, job_id):
-            synced += 1
-    return synced
-
-
-def _decode_archive_row(row: tuple[Any, ...]) -> Dict[str, Any]:
-    keywords_payload = _parse_history_json(row[5])
-    metadata = _parse_history_json(row[6])
-    return {
-        "job_id": row[0],
-        "title": row[1],
-        "snippet": row[2],
-        "source_kind": row[3],
-        "source_value": row[4],
-        "keywords": keywords_payload.get("items", []),
-        "metadata": metadata,
-        "created_at": row[8],
-        "updated_at": row[9],
-    }
-
-
-def get_archive_entry(db_path: str, job_id: str) -> Optional[Dict[str, Any]]:
-    """Fetch a normalized archive entry with archived history payloads."""
-    conn = _connect(db_path)
-    try:
-        cursor = conn.cursor()
-        cursor.execute(
-            """
-            SELECT
-                a.job_id,
-                a.title,
-                a.snippet,
-                a.source_kind,
-                a.source_value,
-                a.keywords,
-                a.metadata,
-                a.search_text,
-                a.created_at,
-                a.updated_at,
-                h.params,
-                h.result,
-                h.status
-            FROM archive_entries AS a
-            LEFT JOIN history AS h ON h.id = a.job_id
-            WHERE a.job_id = ?
-            """,
-            (job_id,),
-        )
-        row = cursor.fetchone()
-    finally:
-        conn.close()
-
-    if row is None:
-        synced = sync_archive_entry_from_history(db_path, job_id)
-        if synced is None:
-            return None
-        return get_archive_entry(db_path, job_id)
-
-    archive_entry = _decode_archive_row(row[:10])
-    archive_entry["params"] = _parse_history_json(row[10])
-    archive_entry["result"] = _parse_history_json(row[11])
-    archive_entry["status"] = row[12]
-    return archive_entry
-
-
-def search_archive_entries(
-    db_path: str,
-    *,
-    query: str | None = None,
-    limit: int = 20,
-) -> list[Dict[str, Any]]:
-    """Search archive entries by source, title, and archived newsletter content."""
-    normalized_limit = max(1, int(limit))
-    sync_archive_entries_from_history(db_path, limit=max(50, normalized_limit * 5))
-
-    conn = _connect(db_path)
-    try:
-        cursor = conn.cursor()
-        cursor.execute(
-            """
-            SELECT
-                job_id,
-                title,
-                snippet,
-                source_kind,
-                source_value,
-                keywords,
-                metadata,
-                search_text,
-                created_at,
-                updated_at
-            FROM archive_entries
-            ORDER BY created_at DESC, job_id DESC
-            LIMIT ?
-            """,
-            (max(50, normalized_limit * 5),),
-        )
-        rows = cursor.fetchall()
-        entries = [_decode_archive_row(row) for row in rows]
-        normalized_query = (query or "").strip().lower()
-        if not normalized_query:
-            return entries[:normalized_limit]
-
-        terms = [term for term in re.split(r"\s+", normalized_query) if term]
-        filtered_entries = [
-            entry
-            for entry, row in zip(entries, rows)
-            if all(term in str(row[7]).lower() for term in terms)
-        ]
-        return filtered_entries[:normalized_limit]
-    finally:
-        conn.close()
-
-
-def record_analytics_event(
-    db_path: str,
-    event_type: str,
-    *,
-    job_id: str | None = None,
-    schedule_id: str | None = None,
-    status: str | None = None,
-    deduplicated: bool = False,
-    duration_seconds: float | None = None,
-    cost_usd: float | None = None,
-    payload: Dict[str, Any] | None = None,
-    occurred_at: str | datetime | None = None,
-) -> str:
-    """Persist a structured analytics event for later aggregation."""
-    event_id = f"evt_{uuid.uuid4().hex}"
-    conn = _connect(db_path)
-    try:
-        cursor = conn.cursor()
-        cursor.execute(
-            """
-            INSERT INTO analytics_events (
-                id,
-                event_type,
-                job_id,
-                schedule_id,
-                status,
-                deduplicated,
-                duration_seconds,
-                cost_usd,
-                payload,
-                created_at
-            )
-            VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?)
-            """,
-            (
-                event_id,
-                event_type,
-                job_id,
-                schedule_id,
-                status,
-                int(deduplicated),
-                duration_seconds,
-                cost_usd,
-                None if payload is None else canonical_json(payload),
-                _serialize_event_timestamp(occurred_at),
-            ),
-        )
-        conn.commit()
-        return event_id
-    finally:
-        conn.close()
-
-
-def list_analytics_events(
-    db_path: str,
-    *,
-    limit: int = 100,
-    event_type_prefix: str | None = None,
-    created_since: str | None = None,
-) -> list[Dict[str, Any]]:
-    """Return recent analytics events for tests and dashboard routes."""
-    conn = _connect(db_path)
-    try:
-        cursor = conn.cursor()
-        if event_type_prefix and created_since:
-            cursor.execute(
-                """
-                SELECT
-                    id,
-                    event_type,
-                    job_id,
-                    schedule_id,
-                    status,
-                    deduplicated,
-                    duration_seconds,
-                    cost_usd,
-                    payload,
-                    created_at
-                FROM analytics_events
-                WHERE event_type LIKE ? AND created_at >= ?
-                ORDER BY created_at DESC, id DESC
-                LIMIT ?
-                """,
-                (f"{event_type_prefix}%", created_since, limit),
-            )
-        elif event_type_prefix:
-            cursor.execute(
-                """
-                SELECT
-                    id,
-                    event_type,
-                    job_id,
-                    schedule_id,
-                    status,
-                    deduplicated,
-                    duration_seconds,
-                    cost_usd,
-                    payload,
-                    created_at
-                FROM analytics_events
-                WHERE event_type LIKE ?
-                ORDER BY created_at DESC, id DESC
-                LIMIT ?
-                """,
-                (f"{event_type_prefix}%", limit),
-            )
-        elif created_since:
-            cursor.execute(
-                """
-                SELECT
-                    id,
-                    event_type,
-                    job_id,
-                    schedule_id,
-                    status,
-                    deduplicated,
-                    duration_seconds,
-                    cost_usd,
-                    payload,
-                    created_at
-                FROM analytics_events
-                WHERE created_at >= ?
-                ORDER BY created_at DESC, id DESC
-                LIMIT ?
-                """,
-                (created_since, limit),
-            )
-        else:
-            cursor.execute(
-                """
-                SELECT
-                    id,
-                    event_type,
-                    job_id,
-                    schedule_id,
-                    status,
-                    deduplicated,
-                    duration_seconds,
-                    cost_usd,
-                    payload,
-                    created_at
-                FROM analytics_events
-                ORDER BY created_at DESC, id DESC
-                LIMIT ?
-                """,
-                (limit,),
-            )
-        rows = cursor.fetchall()
-        return [
-            {
-                "id": row[0],
-                "event_type": row[1],
-                "job_id": row[2],
-                "schedule_id": row[3],
-                "status": row[4],
-                "deduplicated": bool(row[5]),
-                "duration_seconds": row[6],
-                "cost_usd": row[7],
-                "payload": json.loads(row[8]) if row[8] else None,
-                "created_at": row[9],
-            }
-            for row in rows
-        ]
-    finally:
-        conn.close()
-
-
-def get_analytics_dashboard_data(
-    db_path: str,
-    *,
-    window_days: int = 7,
-    recent_limit: int = 25,
-) -> Dict[str, Any]:
-    """Return aggregated analytics metrics and recent events for the dashboard."""
-    since = (
-        (datetime.now(timezone.utc) - timedelta(days=max(1, int(window_days))))
-        .isoformat()
-        .replace("+00:00", "Z")
-    )
-    conn = _connect(db_path)
-    try:
-        cursor = conn.cursor()
-        cursor.execute(
-            """
-            SELECT
-                SUM(CASE WHEN event_type = 'generation.started' THEN 1 ELSE 0 END),
-                SUM(CASE WHEN event_type = 'generation.completed' THEN 1 ELSE 0 END),
-                SUM(CASE WHEN event_type = 'generation.failed' THEN 1 ELSE 0 END),
-                AVG(CASE WHEN event_type = 'generation.completed' THEN duration_seconds END),
-                SUM(CASE WHEN event_type = 'generation.completed' THEN COALESCE(cost_usd, 0) ELSE 0 END)
-            FROM analytics_events
-            WHERE created_at >= ?
-            """,
-            (since,),
-        )
-        generation_row = cursor.fetchone() or (0, 0, 0, None, 0.0)
-
-        cursor.execute(
-            """
-            SELECT
-                SUM(CASE WHEN event_type = 'email.sent' THEN 1 ELSE 0 END),
-                SUM(CASE WHEN event_type = 'email.deduplicated' THEN 1 ELSE 0 END),
-                SUM(CASE WHEN event_type = 'email.failed' THEN 1 ELSE 0 END)
-            FROM analytics_events
-            WHERE created_at >= ?
-            """,
-            (since,),
-        )
-        email_row = cursor.fetchone() or (0, 0, 0)
-
-        cursor.execute(
-            """
-            SELECT
-                SUM(CASE WHEN event_type = 'schedule.created' THEN 1 ELSE 0 END),
-                SUM(CASE WHEN event_type IN ('schedule.execute.enqueued', 'schedule.run_now.queued') THEN 1 ELSE 0 END),
-                SUM(CASE WHEN event_type IN ('schedule.execute.completed', 'schedule.run_now.completed') THEN 1 ELSE 0 END),
-                SUM(CASE WHEN event_type IN ('schedule.execute.failed', 'schedule.run_now.failed') THEN 1 ELSE 0 END)
-            FROM analytics_events
-            WHERE created_at >= ?
-            """,
-            (since,),
-        )
-        schedule_row = cursor.fetchone() or (0, 0, 0, 0)
-    finally:
-        conn.close()
-
-    generation_started = int(generation_row[0] or 0)
-    generation_completed = int(generation_row[1] or 0)
-    generation_failed = int(generation_row[2] or 0)
-    success_base = generation_completed + generation_failed
-    success_rate = (
-        round((generation_completed / success_base) * 100, 1) if success_base else None
-    )
-
-    return {
-        "window_days": max(1, int(window_days)),
-        "summary": {
-            "generation": {
-                "started": generation_started,
-                "completed": generation_completed,
-                "failed": generation_failed,
-                "success_rate": success_rate,
-                "average_duration_seconds": (
-                    round(float(generation_row[3]), 2)
-                    if generation_row[3] is not None
-                    else None
-                ),
-                "total_cost_usd": round(float(generation_row[4] or 0.0), 4),
-            },
-            "email": {
-                "sent": int(email_row[0] or 0),
-                "deduplicated": int(email_row[1] or 0),
-                "failed": int(email_row[2] or 0),
-            },
-            "schedule": {
-                "created": int(schedule_row[0] or 0),
-                "queued": int(schedule_row[1] or 0),
-                "completed": int(schedule_row[2] or 0),
-                "failed": int(schedule_row[3] or 0),
-            },
-        },
-        "recent_events": list_analytics_events(
-            db_path,
-            limit=max(1, int(recent_limit)),
-            created_since=since,
-        ),
-    }
